Move extractDisplayFields and PerviewData to new files
