+++ clear-signing-developer-tools/src/app/page.tsx
@@ -1,23 +1,20 @@
 import { DevicesDemo } from "~/app/DevicesDemo";
 import { PreviewForm } from "~/app/PreviewForm";
-import { type PreviewData } from "~/app/raw-data-example/page";
+import { type PreviewData } from "~/types/PreviewData";
 import { UI } from "~/app/UI";
+import { extractDisplayFields } from "~/app/raw-data-example/extractDisplayFields";
+import poapMetaDataFile from "../../../registry/poap/calldata-PoapBridge.json";
+import { type ERC7730Schema } from "~/types";
 
 export default function HomePage() {
-  const mockData = {
-    metadata: {
-      owner: "POAP",
-      info: {
-        legalName: "Proof of Attendance Protocol",
-        lastUpdate: "2021-09-01",
-        url: "https://poap.xyz",
-      },
-    },
-    intent: "Mint POAP",
-    displays: [{ label: "tx1", displayValue: "0xbe936…e403" }],
-  } satisfies PreviewData;
+  const wipParsedData = extractDisplayFields(
+    poapMetaDataFile as unknown as ERC7730Schema,
+  );
 
-  const data = mockData;
+  const data = {
+    ...wipParsedData,
+    intent: "Mint POAP",
+  } satisfies PreviewData;
 
   return (
     <main>
@@ -26,7 +23,7 @@ export default function HomePage() {
         <PreviewForm />
       </div>
       <DevicesDemo data={data} />
-      <pre className="container p-16">{JSON.stringify(mockData, null, 2)}</pre>
+      <pre className="container p-16">{JSON.stringify(data, null, 2)}</pre>
     </main>
   );
 }
