+++ clear-signing-developer-tools/src/types/PreviewData.d.ts
@@ -1,0 +1,17 @@
+export type PreviewData = {
+  intent: string;
+  metadata: {
+    owner: string;
+    info: {
+      legalName: string;
+      lastUpdate: string;
+      url: string;
+    };
+  };
+  displays: DisplayItem[];
+};
+
+interface DisplayItem {
+  label: string;
+  displayValue: string;
+}
